Pressure correction factor for dissolved oxygen measurements

Pressure correction factor for dissolved oxygen measurements

Usage

ss_dissolved_oxygen_pressure_correction(dat_wide, sal = NULL, p_atm = NULL)

Arguments

dat_wide

Data frame with at least one column: temperature_degree_C. Corresponding pressure (atm) data may be included in column pressure_atm. Additional columns will be ignored and returned.

sal

A single value of salinity (psu). This value must be specified if there is no salinity_psu column in dat_wide. Default is sal = NULL. Note: if sal is specified when there is a salinity_psu column in dat_wide, the function will stop with an error.

p_atm

A single value of barometric pressure (atm). This value should be specified if there is no Pressure column in dat_wide. Default is p_atm = NULL. Note: if p_atm is specified when there is a pressure_atm column in dat_wide, function will stop with an error.

Value

Returns dat_wide with additional column(s), F_p and

pressure_atm.

Details

Dissolved oxygen concentration (and partial pressure?) measurements should be corrected for atmospheric pressure if the pressure deviates from substantially from 1 atm (Benson & Krause, 1984).

This function calculates pressure correction factor following Benson and Krause (1984), as suggested by USGS (2011). This is similar to what is described in the HOBO manual HOBO U26 Percent Saturation Calculation.pdf (except this function accounts for salinity in the water vapour calculation).

Equation 24 in Benson & Krause 1984:

$$C_{p} = C_{star} * Pressure * (((1 - P_{wv} / Pressure) * (1 - theta * Pressure)) / ((1 - P_{wv}) * (1 - theta)))$$

$$F_{p} = Pressure * (((1 - P_{wv} / Pressure) * (1 - theta * Pressure)) / ((1 - P_{wv}) * (1 - theta)))$$

\(P_{wv}\) is water vapour pressure, which depends on temperature and salinity and \(theta\) depends on the second virial coefficient of oxygen, and is calculated using temperature (see Table 2 of Benson and Krause 1984).

References

Benson, Bruce B., Krause, Daniel, (1984), The concentration and isotopic fractionation of oxygen dissolved in freshwater and seawater in equilibrium with the atmosphere, Limnology and Oceanography, 3, doi: 10.4319/lo.1984.29.3.0620.

USGS. Change to Solubility Equations for Oxygen in Water. Technical Memorandum 2011.03. USGS Office of Water Quality, 2011.

See also