Plot variables at depth

Plot variables at depth

Usage

ss_ggplot_variables(
  dat,
  superchill = NULL,
  color_palette = NULL,
  legend_name = "Depth (m)",
  legend_position = "right",
  axis_label_newline = TRUE,
  point_size = 0.25
)

Arguments

dat

Data frame of sensor string data in wide or long format.

superchill

Logical argument indicating whether to adding shading to indicate temperatures below the superchill threshold (<= - 0.7 degrees C). If superchill = NULL (the default), shading will be applied if any temperature values are less than or equal to the threshold.

color_palette

Optional vector of hex colors onto which depth will be mapped.

legend_name

Name for the depth legend. Default is legend_name = "Depth (m)".

legend_position

Position for the depth legend. Default is legend.position = "right".

axis_label_newline

Logical argument indicating whether to put units on a new line.

point_size

Numeric value indicating size of points.

Value

Returns a ggplot object of ocean variables plotted over time and coloured by sensor depth.