Extract information from deployment log

log must include the following columns (or the information will be recorded as NA):

Usage

ss_parse_log(
  log,
  deployment_dates = TRUE,
  area_info = TRUE,
  sn_table = TRUE,
  config = TRUE
)

Arguments

log

Log data frame, e.g., as exported from ss_read_log_simple().

deployment_dates

Logical argument indicating whether to parse and return the deployment and retrieval dates.

area_info

Logical argument indicating whether to parse and return the location details.

sn_table

Logical argument indicating whether to parse and return the serial numbers and sensor depth.

config

Logical argument indicating whether to parse and return the string configuration.

Value

Returns a list with up to 4 elements. deployment_dates is a data frame with two columns: start_date (the date of deployment) and end_date (date of retrieval). area_info is a data frame with five columns: county, waterbody, latitude, longitude, station, and lease. sn_table is a data frame with three columns: log_sensor (sensor name as recorded in the log), sensor_serial_number, and depth (sensor depth below the surface at low tide, from the Sensor_Depth column). string_configuration is a character string indicating how the sensor string was moored (i.e., whether the sensors float with the tide).

Details

county: county where string was deployed

waterbody: waterbody where string was deployed

station: the station name

lease: the aquaculture lease number, if applicable (NA otherwise)

deployment_date: deployment date, in the order "YYYY-MM-DD"

retrieval_date: retrieval date, in the order "YYYY-MM-DD"

deployment_latitude: The latitude at which the string was deployed

deployment_longitude The longitude at which the string was deployed (must be a negative value)

sensor_type The type of sensor

sensor_serial_number The sensor serial number

sensor_depth_m: Depth at which the sensor was deployed, in m

All other columns will be ignored.

Entries in the sensor_type column must include the string "aquameasure", "hobo", "tidbit", or "vr2ar" (not case sensitive).

The function will stop with an Error if there is more than one eligible file (csv, .xlsx or .xls) in the Log folder.

A warning is printed if deployment_longitude is a positive value.

A Warning message is printed when the function does not recognize a sensor in the log.

A message is printed to the console if there is more than one unique entry in waterbody, station, deployment_date, retrieval_date, deployment_latitude, or deployment_longitude.